requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
anyio==3.6.2
astunparse==1.6.3
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.2
flatbuffers==23.5.9
gast==0.4.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.55.0
h11==0.14.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
jax==0.4.10
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
typing-extensions==4.6.0
urllib3
uvicorn==0.22.0
Werkzeug==2.3.4
wrapt==1.14.1
zipp==3.15.0
